NewAcademia/sensus-lc-web-test/src/main/java/com/sensus/lc/controller/tags/unittest/TagViewControllerTest.java
package com.sensus.lc.controller.tags.unittest;

import static org.hamcrest.Matchers.containsString;
import static org.hamcrest.Matchers.notNullValue;
import static org.springframework.test.web.server.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.server.result.MockMvcResultMatchers.model;
import static org.springframework.test.web.server.result.MockMvcResultMatchers.status;

import org.junit.Test;
import org.springframework.web.servlet.ModelAndView;

import com.sensus.common.util.SensusAppContext;
import com.sensus.lc.controller.tags.TagViewController;
import com.sensus.lc.controller.util.AbstractTestBase;
import com.sensus.lc.controller.util.SessionAuthenticationTestUtil;
import com.sensus.lc.controller.util.enums.ModeEnum;
import com.sensus.lc.controller.util.enums.TestMessageEnum;

public class TagViewControllerTest extends AbstractTestBase
{

	/** The Constant TAG_MAIN. */
	private static final String TAG_MAIN = "/tag/tag_main";

	/** The Constant TAGS. */
	private static final String TAGS = "tags";

	/** The fetch all. */
	private final String FETCH_LIST = "/tag";

	/** The Constant CONTROLLER_EXCEPTION_MSG. */
	public static final String CONTROLLER_EXCEPTION_MSG = "TagViewController";

	/**
	 * Fetch list.
	 * 
	 * @throws Exception the exception
	 */
	@Test
	public void fetchList() throws Exception
	{

		// Success situation
		getTagMock().setMode(ModeEnum.MODE_SUCCESS);
		getMockMvc().perform(
				get(FETCH_LIST)
						.session(SessionAuthenticationTestUtil.getSessionTest()))
				.andExpect(status().isOk())
				// TODO Change to check for response
				.andExpect(model().attribute(TAGS, notNullValue()))
				.equals(new ModelAndView(TAG_MAIN, TAGS, new String()));

		// Failure situation
		getTagMock().setMode(ModeEnum.MODE_FAILURE);
		getMockMvc().perform(
				get(FETCH_LIST)
						.session(SessionAuthenticationTestUtil.getSessionTest()))
				.andExpect(status().isOk())
				.andExpect(model().attribute(TAGS, containsString(TestMessageEnum.MESSAGE_ERROR.getValue())))
				.equals(new ModelAndView(TAG_MAIN, TAGS, new String()));
	}

	/**
	 * Gets the tag mock.
	 * 
	 * @return the tag mock
	 */
	private TagBCFMockImpl getTagMock()
	{
		return (TagBCFMockImpl)SensusAppContext.getApplicationContext()
				.getBean(TagViewController.class).getTagBCF();
	}
}